move RoomMembers to own dir
